Refactor validation process: extract validation logic into separate files and improve error handling
